Refactored code to match React style
